feat: added RequestValidator tests and created the Request folder
