+++ includes/classes/RequestValidator.hpp
@@ -16,11 +16,16 @@ class RequestValidator
 		~RequestValidator();
 
 		RequestValidator    		&request_validator( Conf &conf, Request &request);
+		HttpMethodEnum::httpMethod	method_validator(Request& request);
+		void 						path_validator(Conf& conf, Request& request);
+		void						body_validator(Request& request);
+		void						http_version_validator(Request& request);
 		HttpMethodEnum::httpMethod	get_method(void) const;
 		bool				        get_path(void) const;
 		bool				        get_http_version(void) const;
 		bool				        get_body(void) const;
 		bool				        get_header(std::string header_name) const;
+
 
 
 		HttpMethodEnum::httpMethod      _method;
